The dictionary to be saved or loaded can now be any dictionary.
The dictionary to be saved or loaded must now be transferred when the save or load function is called. The previous “data” dictionary therefore no longer exists.
Thanks to this change, it is now possible to save or load individual parts so that not everything always has to be saved or loaded.

If you want to save an empty dictionary, you have to set the variable can_save_empty_data, which is at the top, to true. Otherwise, empty dictionaries will not be saved.