Added a queue for the save function
I have added a queue to the save function, in which the save requests are stored, which are requested during a running save process. When the current save process is terminated, a check is made to see if there is one waiting in the queue. If this is the case, the next save request from the queue is processed.