Impr. loading queue. Added a cleanup for old data. Added some comments
Added a queue to the loading system. This allows a new loading process to be started as soon as the previous one is completed if a new one is triggered during an ongoing loading process.
If several loading processes are added to the queue, but they have the same name, the new process replaces the old one.

Added the delete_old_loaded_data variable. If this variable is set to true, old data from the loaded data,which no longer exist in the dictionary into which they are to be loaded, gets deleted from the save file.

Added some comments to make the script simpler to understand.