Impr. save and loading queue
The save function now call at the end the check_save_queue function to check, if another save should happen. If there is another the save_data function is called, with the information it needs for a save process.
The same happens in the loading function. This uses the check_loading_queue.