Impr. queues, signal connection, Process status
The queues now have a common function for checking instead of each having a separate one.
The signals are now connected via a loop. This means that each one does not have to be written individually to the ready function. The name of a new signal only needs to be added to the “SIGNALS” array.
The process status is now displayed via the process_status variables. These receive their status via the assignment of a status from the enum process_status.