Impr. new log_message and file_accessfunction
Combined the log_debug and the log_error function to a log_message function. It can output both DEBUG and ERROR messages.

Added the file_access function which returns the currently required file access. The corresponding access depends on whether the data is to be saved or loaded and whether it is encrypted or not.